(GH-22) Add itch.io embed shortcode
This commit adds a new shortcode, itchio, to the theme.

This shortcode takes an id value for the embeddable widget
and creates an iframe to host it in the platen site.

The shortcode takes parameters for square, dark mode, and
whether or not to include a linkback to the project page.

Resolves #22