(GH-35) Inject CSS into body of _default/baseof.html
This commit modifies the `_default/baseof.html` template to inject
a CSS class into the `body` tag, enabling Platen site maintainers
to override the styling for any page(s) via this mechanism.

This is a patch on top of the `layouts/_default/baseof.html` file
from the usptream project (alex-shpak/hugo-book). If the upstream
file is updated, we will need to replace our copy of the template
with the newer version and then cherry-pick this commit on top.