(MAINT) Fix CSS override partial
Prior to this commit, the `getCssClass` partial that provides the CSS
override functionality on a per-page and per-section basis was broken
if a site had any duplicate section names, even if they have different
ancestors.

This change fixes the bug by introducing logic to search for the fully
qualified section instead of the section name alone. For example,
instead of `concepts`, it searches for `/modules/platen/concepts`.